👔 Create row for user if not exists when opening doors
  Introduce logic to check if a row with the user's ID exists in the OpenedDoors table. If not, create a new row for the user before adding the opened doors. Enhances the functionality to ensure proper data handling.

+++ advent_calendar_backend/src/main/java/de/iks/advent_calendar_backend/rest/service/AuthService.java
@@ -29,7 +29,7 @@ public class AuthService {
 				.map(GrantedAuthority::getAuthority)
 				.toList();
 
-		var token = jwtIssuer.issue(principal.getUserId(), principal.getUsername(), roles);
+		var token = jwtIssuer.issue((long) principal.getUserId(), principal.getUsername(), roles);
 
 		ResponseCookie cookie = ResponseCookie.from("accessToken", token)
 			.httpOnly(true)
