+++ advent_calendar_backend/src/main/java/de/iks/advent_calendar_backend/security/UserPrincipal.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Builder
 public class UserPrincipal implements UserDetails {
 
-	private final Long userId;
+	private final int userId;
 	private final String username;
 	@JsonIgnore
 	private final String password;
